+++ Appointment/appoint/manage.py
@@ -11,7 +11,7 @@ from Appointment.appoint.jobs import set_scheduler, cancel_scheduler
 from Appointment.extern.wechat import MessageType, notify_appoint
 from Appointment.extern.jobs import set_appoint_reminder
 from utils.wrap import return_on_except, stringify_to
-from achievement.unlock_api import unlock_achievement
+from achievement.api import unlock_achievement
 from generic.models import User
